Reorganized majorinfo.py. Collects major links before individually scrapping
